+++ .gitignore
@@ -2,6 +2,7 @@
 .cache
 node_modules
 yarn-error.log
+package-lock.json
 yarn.lock
 
 # Build directory
