+++ gatsby-config.js
@@ -2,11 +2,11 @@ module.exports = {
   siteMetadata: {
     title: 'Kyle Ross - Web Developer',
     author: 'Kyle Ross',
-    siteUrl: 'https://badger3000.com',
-    description:
-      'Kyle Ross is a front end developer, based in the Prescott, Arizona',
+    description: 'Personal site of Kyle Ross',
   },
   plugins: [
+    'gatsby-plugin-preact',
+    'gatsby-plugin-react-helmet',
     {
       resolve: `gatsby-source-filesystem`,
       options: {
@@ -14,6 +14,7 @@ module.exports = {
         path: `${__dirname}/src/assets/images/`,
       },
     },
+    'gatsby-transformer-sharp',
     'gatsby-plugin-sharp',
     {
       resolve: `gatsby-plugin-manifest`,
@@ -32,8 +33,6 @@ module.exports = {
         orientation: 'portrait', // This path is relative to the root of the site.
       },
     },
-    'gatsby-transformer-sharp',
-    `gatsby-plugin-image`,
     {
       resolve: 'gatsby-plugin-preconnect',
       options: {
@@ -47,7 +46,8 @@ module.exports = {
         ],
       },
     },
-
+    'gatsby-plugin-preload-link-crossorigin',
     'gatsby-plugin-sass',
+    'gatsby-plugin-offline',
   ],
 }
