+++ gatsby-ssr.js
@@ -5,6 +5,3 @@
  */
 
 // You can delete this file if you're not using it
-export const onRenderBody = ({ setHtmlAttributes }) => {
-  setHtmlAttributes({ lang: 'en-GB' })
-}
