+++ package.json
@@ -1,25 +1,29 @@
 {
   "name": "badger3000",
   "description": "Gatsby site",
-  "version": "1.1.0",
+  "version": "1.0.0",
   "author": "Kyle Ross",
   "dependencies": {
-    "gatsby": "^5.4.2",
-    "gatsby-background-image": "^1.6.0",
-    "gatsby-plugin-image": "^3.4.0",
-    "gatsby-plugin-manifest": "^5.4.0",
-    "gatsby-plugin-offline": "^6.4.0",
+    "gatsby": "^2.24.66",
+    "gatsby-background-image": "^1.1.2",
+    "gatsby-image": "^2.4.20",
+    "gatsby-plugin-manifest": "^2.4.32",
+    "gatsby-plugin-offline": "^3.2.29",
+    "gatsby-plugin-preact": "^4.0.13",
     "gatsby-plugin-preconnect": "^1.1.54",
-    "gatsby-plugin-sass": "^6.4.0",
-    "gatsby-plugin-sharp": "^5.4.0",
-    "gatsby-source-filesystem": "^5.4.0",
-    "gatsby-transformer-sharp": "^5.4.0",
-    "gbimage-bridge": "^0.2.2",
-    "intersection-observer": "^0.12.2",
-    "node-sass": "^8.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-images": "^0.6.7"
+    "gatsby-plugin-preload-link-crossorigin": "^1.0.2",
+    "gatsby-plugin-react-helmet": "^3.3.12",
+    "gatsby-plugin-sass": "^2.3.14",
+    "gatsby-plugin-sharp": "^2.6.37",
+    "gatsby-source-filesystem": "^2.3.31",
+    "gatsby-transformer-sharp": "^2.5.16",
+    "node-sass": "^4.13.1",
+    "preact": "^10.5.3",
+    "preact-render-to-string": "^5.1.10",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^5.2.1",
+    "react-images": "^1.0.0"
   },
   "keywords": [
     "gatsby"
@@ -28,7 +32,6 @@
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
-    "debug": "gatsby build && gatsby serve",
     "format": "prettier --write \"**/*.js\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
