+++ src/assets/scss/main.scss
@@ -976,14 +976,11 @@ input[type='button'],
   > * {
     @include vendor('flex-shrink', '0');
     width: 100%;
-    position: relative;
-    z-index: 101;
   }
 
   > .inner {
     @include vendor('flex-grow', '1');
     margin: 0 0 ($size-header-pad * 0.5) 0;
-    
   }
 
   strong,
@@ -1009,18 +1006,6 @@ input[type='button'],
   .image.avatar {
     margin: 0 0 (_size(element-margin) * 0.5) 0;
     width: 6.25em;
-  }
-  &:before {
-    background-color: rgba(0, 0, 0, 0.2);
-    opacity: 1;
-    display: inline-block;
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top:0;
-    left: 0;
-    z-index: 100;
-    ;
   }
 }
