+++ src/components/Gallery/constants/defaultImages.js
