+++ src/components/Gallery/index.js
