+++ src/components/Image.js
@@ -1,0 +1,47 @@
+import React from 'react'
+import { StaticQuery, graphql } from 'gatsby'
+import Img from 'gatsby-image'
+
+// Note: You can change "images" to whatever you'd like.
+
+const Image = (props) => (
+  <StaticQuery
+    query={graphql`
+      query {
+        images: allFile {
+          edges {
+            node {
+              relativePath
+              name
+              childImageSharp {
+                fluid(maxWidth: 600) {
+                  ...GatsbyImageSharpFluid_withWebp
+                }
+              }
+            }
+          }
+        }
+      }
+    `}
+    render={(data) => {
+      const image = data.images.edges.find((n) => {
+        return n.node.relativePath.includes(props.filename)
+      })
+      if (!image) {
+        return null
+      }
+
+      //const imageSizes = image.node.childImageSharp.sizes; sizes={imageSizes}
+      return (
+        <Img
+          loading="eager"
+          alt={props.alt}
+          fluid={image.node.childImageSharp.fluid}
+          style={{ position: 'inhert' }}
+        />
+      )
+    }}
+  />
+)
+
+export default Image
