--- src/components/hooks/site-meta.js
@@ -1,19 +1,0 @@
-import { useStaticQuery, graphql } from 'gatsby'
-
-export const useSiteMetadata = () => {
-  const { site } = useStaticQuery(
-    graphql`
-      query SiteMetaData {
-        site {
-          siteMetadata {
-            title
-            author
-            siteUrl
-            description
-          }
-        }
-      }
-    `
-  )
-  return site.siteMetadata
-}
