+++ src/components/layout.js
@@ -4,10 +4,10 @@ import '../assets/scss/main.scss'
 import Header from './Header'
 
 const Template = ({ children }) => (
-  <>
+  <div>
     <Header />
     {children}
-  </>
+  </div>
 )
 
 export default Template
