+++ src/pages/confirmation.js
@@ -1,10 +1,16 @@
 import React from 'react'
+import Helmet from 'react-helmet'
 import Layout from '../components/layout'
 
-import { useSiteMetadata } from '../components/hooks/site-meta'
 const NotFoundPage = () => {
+  const siteTitle = 'Thank you for contacting me'
+  const siteDescription = 'Your submission is confirmed'
   return (
     <Layout>
+      <Helmet>
+        <title>{siteTitle}</title>
+        <meta name="description" content={siteDescription} />
+      </Helmet>
       <div id="main">
         <h1>Thanks for Contacting Me</h1>
         <p>Will be in touch soon</p>
@@ -17,14 +23,3 @@ const NotFoundPage = () => {
 }
 
 export default NotFoundPage
-
-export function Head() {
-  const { title, description } = useSiteMetadata
-
-  return (
-    <>
-      <title>{title}</title>
-      <meta name="description" content={description} />
-    </>
-  )
-}
